Simplify solution for SwiftUI 5
