+++ Sources/OnRemoveFromParent/Common/View+Identify.swift
@@ -7,28 +7,22 @@
 
 import SwiftUI
 
-private struct ViewIdentityKey: EnvironmentKey {
-  static let defaultValue: AnyHashableSendable? = nil
-}
-
-extension EnvironmentValues {
-  var viewIdentity: AnyHashableSendable? {
-    get { self[ViewIdentityKey.self] }
-    set { self[ViewIdentityKey.self] = newValue }
-  }
-}
-
 extension View {
   /// Binds a view's identity to the given proxy value.
   ///
   /// When the proxy value specified by the `id` parameter changes, the
   /// identity of the view — for example, its state — is reset.
   @inline(__always)
+  @ViewBuilder
   public func identify<ID: Hashable & Sendable>(
     as id: ID
   ) -> some View {
-    self
-      .environment(\.viewIdentity, AnyHashableSendable(id))
-      .id(id)
+    if #available(iOS 17, macCatalyst 17, macOS 14, tvOS 17, visionOS 2, watchOS 11, *) {
+      self.id(id)
+    } else {
+      self.modifier(
+        IdentifyViewModifier(id: id)
+      )
+    }
   }
 }
